requirements.txt
aiogram==3.18.0
pydantic-settings>=2.0.0
groq>=0.11.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
python-docx>=1.1.0
